apps/model-control-plane/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
